backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
PyMySQL>=1.1.0
alembic>=1.13.0
groq>=0.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 